Updated CartStorageService and added ICartStorageService interface
Implemented the `ICartStorageService` interface in `CartStorageService.cs` and updated the `List`, `AddItem`, and `Delete` methods to work with a new private dictionary. Added a project reference to `Amazon.Payment.Sql.csproj` in `Amazon.Payment.Api.csproj`. Updated `CartController.cs` to use `ICartStorageService` and modified the `AddItem` and `Delete` methods to take additional parameters. Registered `ICartStorageService` as a singleton service in `Program.cs`. Added new `ICartStorageService.cs` file that defines the interface and a new `CartStorageService2.cs` file for a second implementation of the interface, currently throwing `NotImplementedException`.