Added CartId property and Docker support
A new `CartId` property has been added to the `Item` class and is now used in the `AddItem` and `List` methods of `CartStorageServiceWithSql.cs`. This property has also been added to the model snapshot and a new migration has been created to add the `CartId` column to the `Items` table in the database. Docker support has been added to the project, including Docker-related settings in `launchSettings.json` and `Amazon.Payment.Api.csproj`, a new `.dockerignore` file, and a new Dockerfile. User secrets have also been enabled for the project.